requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
mangum==0.12.3
sqlalchemy==1.4.27
pandas==1.3.4
boto3==1.20.12
pymysql==1.0.2
scikit-learn==1.0.1
singleton-decorator==1.0.0
pydantic==1.8.2
passlib==1.7.4
bcrypt==3.2.0
PyJWT==2.3.0